Adding example that shows each term's contribution to covariance in notebook.
PR changes
